# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Machanic Shop API"
  description: "API to manage a mechanic shop."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: Customer login (generate token)
      description: Post login credentials (email and password), if valid a user specific token will be minted and returned.
      parameters:
        - in: body
          name: body
          description: Login Credentials
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        201:
          description: Successful login
          schema:
            type: object
            properties:
              token:
                type: string
          examples:
            application/json:
              message: "Successfully Logged In"
              status: "success"
              token: "eyJhbGciOiJI...."

#create customer
  /customers:
    post:
      tags:
        - Customers
      summary: Create a Customer
      description: Send a JSON payload with customer information to create a customer in the database.
      parameters:
        - in: body
          name: body
          description: Information to be created
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        201:
          description: Successful creation with hashed password
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"

#upadte cutomer
    put:
      tags:
        - Customers
      summary: Update a Customer (Token Required)
      description: Update an existing Customer, requires that customer to be logged in.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Information used to update customer
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        201:
          description: Successful update of customer
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"

#get all customers
    get:
      tags:
        - Customers
      summary: Get all Customers
      description: Endpoint to return all Customers
      responses:
        200:
          description: Successful retrieval
          schema:
            $ref: '#/definitions/CustomerListResponse'
          examples:
            application/json:
              [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
              }]

#delete customer
    delete:
      tags:
        - Customers
      summary: Delete a Customer (Token Required)
      description: Delete an existing Customer, requires customer to be logged in.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful deletion of Customer
          examples:
            application/json:
              message: "Customer <customer.name> was deleted."
            
#get single customer
  /customers/{id}:
    get:
      tags:
        - Customers
      summary: Get a specific Customers using ID
      description: Pass in a Customer ID as a path parameter and returns specific Customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
          description: The Customer ID
      responses: 
        200:
          description: Successful query
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              email: "johndoe@email.com"
              id: 1
              name: "John Doe"
              password: "<hashed password>"
              phone: "123-123-1234"

#search customers by name
  /customers/search:
    get:
      tags:
        - Customers
      summary: Get Customer(s) using search query by name
      description: Receives name as a query parameter and returns returns all Customers who's name matches the query parameter passed in.
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Name of Customer(s) to search.
      responses: 
        200:
          description: Successful search
          schema:
            $ref: '#/definitions/SearchCustomerList'
          examples:
            application/json:
              customers: [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
              }]
              
#get all customers (paginated)
  /customers/paginated:
    get:
      tags:
        - Customers
      summary: Get Customer(s) using paginated query parameters
      description: Receives page and per_page as a query parameters and returns the number of Customer(s) as specified by per_page query parameters on the specified page query parameter.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: The page number to display Customer(s)
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
          description: The number of Customers to display per page.
      responses:
        200:
          description: Successful retrieval
          schema:
            $ref: '#/definitions/CustomerListResponse'
          examples:
            application/json:
              [{
                email: "johndoe@email.com",
                id: 1,
                name: "John Doe",
                phone: "123-123-1111",
              },{
                email: "janedoe@email.com",
                id: 2,
                name: "Jane Doe",
                phone: "123-123-2222",
              }]

#get single customer's tickets
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get all tickets for single customer (Token Required)
      description: Retrieve all tickets associated with single customer, requires customer to be logged in.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful retrieval of Customer tickets
          schema:
            $ref: '#/definitions/CustomerTicketsResponse'
          examples:
            application/json:
              "tickets": [
                    {
                        "VIN": "VIN12345",
                        "id": 2,
                        "mechanics": [
                            {
                                "id": 1,
                                "name": "Mechanic1"
                            }
                        ],
                        "serialized_parts": [
                            {
                                "id": 9,
                                "part_description": {
                                    "brand": "Pro Tire's",
                                    "id": 4,
                                    "part_name": "All-season tire",
                                    "price": 100.0
                                }
                            }
                        ],
                        "service_date": "2025-05-08",
                        "service_desc": "Replace Tires",
                        "services": [
                            {
                                "id": 1,
                                "labor_hours": 0.5,
                                "labor_rate": 100.0,
                                "name": "change tires"
                            }
                        ]
                    }]
   
#=================part description paths===============
#create part description
...
definitions: 
...
  CustomerPayload:
    type: object
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
  CustomerResponse:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
  CustomerListResponse: #password excluded
    type: array
    items:
      type: object
      properties:
        email:
          type: string
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
  SearchCustomerList:
      type: object
      properties:
        customers:
          type: array
          items:
            type: object
            properties:
              email:
                type: string
              id:
                type: integer
              name:
                type: string
              phone:
                type: string
...
  CustomerTicketsResponse:
   type: object
   properties:
    tickets:
      type: array
      items:
        type: object
        properties:
          VIN:
            type: string
          id:
            type: integer
          mechanics:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
          serialized_parts:
            type: array
            items:
              type: object
              properties:
                id: 
                  type: integer
                part_description: 
                  type: object
                  properties:
                    brand:
                      type: string
                    id:
                      type: integer
                    part_name:
                      type: string
                    price:
                      type: integer
          service_desc:
            type: string
          service_date:
            type: string
          services:
            type: array
            items: 
              type: object
              properties:
                id:
                  type: integer
                labor_hours:
                  type: integer
                labor_rate:
                  type: integer
                name:
                  type: string



#=========== Mechanic Definitions ==========
...